Moved color functions to headers which can be used stand-alone, or with ArtColors, or RayLib Some function code simplified.  Help button now works.  Color picker has a toggle button.  RYB mixing areas work, but don't go to blend bars without Color Picker to select them.
